Extract the ballot screens into their own modules
tournament-ballot.tsx was 4,222 lines holding ten components, which is why a
schema change surfaced one type error where thirty real problems existed and
why none of its logic could be tested.

Nine components move to components/tournaments/ballot/: the timer, argument
flow, fact checking, collaborative notes, the speaker position manager, the
loading skeleton, the row and card list items, the details dialog and the
flag dialog. SCORING_CATEGORIES, useNames and the debate status helpers move
with them, since several modules share those.

The page keeps JudgingInterface and the shell and is now 1,889 lines. The
code moved verbatim, so this changes structure and not behaviour: the same
191 tests pass, both projects typecheck, and the production build is clean.
