+++ components/tournaments/ballot/ballot-skeleton.tsx
@@ -1,0 +1,116 @@
+"use client";
+
+import React, { useEffect, useMemo, useRef, useState } from "react";
+import { useMutation, usePaginatedQuery, useQuery } from "convex/react";
+import { api } from "@/convex/_generated/api";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+import { Badge } from "@/components/ui/badge";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Textarea } from "@/components/ui/textarea";
+import { Separator } from "@/components/ui/separator";
+import { Alert, AlertDescription } from "@/components/ui/alert";
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible";
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { Skeleton } from "@/components/ui/skeleton";
+import { ScrollArea } from "@/components/ui/scroll-area";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Drawer, DrawerContent, DrawerHeader, DrawerTitle } from "@/components/ui/drawer";
+import { Progress } from "@/components/ui/progress";
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
+import {
+  AlertCircle,
+  AlertTriangle,
+  BarChart3,
+  Brain,
+  CheckCircle,
+  CheckSquare, CircleCheck,
+  Clock,
+  Crown,
+  Download,
+  Edit3,
+  Eye,
+  FileText,
+  Flag,
+  Grid,
+  Link2,
+  List,
+  Loader2,
+  MessageSquare,
+  Mic,
+  MicOff,
+  Minus,
+  Pause,
+  Play,
+  Plus,
+  Search,
+  Send,
+  Shield,
+  Square,
+  Target,
+  Timer,
+  Users2,
+  ChevronDown,
+  ChevronRight
+} from "lucide-react";
+import { toast } from "sonner";
+import { Id } from "@/convex/_generated/dataModel";
+import { useGemini } from "@/hooks/use-gemini";
+import { speechTotal, rangesFor, validateSpeechScore, validateOutcome, type SpeechScore } from "@/lib/scoring/wsdc";
+import { EMPTY_SCORE, MIN_RFD_LENGTH } from "@/components/tournaments/ballot/types";
+import { isSupportedFormat } from "@/lib/tournament-formats";
+import type {
+  ArgumentFlowEntry,
+  BallotRecord,
+  BallotTournament,
+  EnrichedDebate,
+  FactCheckEntry,
+  PositionsBySpeaker,
+  SharedNoteEntry,
+  UserRole,
+} from "@/components/tournaments/ballot/ballot-types";
+import type { BallotSubmission } from "@/components/tournaments/ballot/submission";
+import type { SpeakerPosition } from "@/components/tournaments/ballot/types";
+import { clearDraft, loadDraft, saveDraft } from "@/lib/offline/drafts";
+import { formatClock } from "@/lib/scoring/speech-timing";
+import { cn } from "@/lib/utils";
+import { useOffline } from "@/hooks/use-offline";
+import { Checkbox } from "@/components/ui/checkbox";
+import { useDebounce } from "@/hooks/use-debounce";
+
+export function BallotSkeleton() {
+  return (
+    <div className="space-y-4">
+      <div className="flex flex-col lg:flex-row lg:items-center bg-brown rounded-t-md lg:justify-between gap-4 p-3">
+        <div>
+          <Skeleton className="h-6 w-48 mb-2" />
+          <Skeleton className="h-4 w-64" />
+        </div>
+        <div className="flex items-center gap-3">
+          <Skeleton className="h-8 w-24" />
+          <Skeleton className="h-8 w-18" />
+          <Skeleton className="h-8 w-24" />
+          <Skeleton className="h-8 w-32" />
+          <Skeleton className="h-8 w-8" />
+        </div>
+      </div>
+
+      {Array.from({ length: 6 }).map((_, i) => (
+        <div key={i} className="flex items-center space-x-4 p-4 border rounded-lg">
+          <Skeleton className="h-12 w-12 rounded" />
+          <div className="space-y-2 flex-1">
+            <Skeleton className="h-4 w-[250px]" />
+            <Skeleton className="h-4 w-[200px]" />
+          </div>
+          <div className="flex space-x-2">
+            <Skeleton className="h-8 w-16" />
+            <Skeleton className="h-8 w-16" />
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+}
+
