+++ components/tournaments/tournament-ballot.tsx
@@ -79,6 +79,18 @@ import { useSpeechTimer } from "@/components/tournaments/ballot/use-speech-timer
 import { clearDraft, loadDraft, saveDraft } from "@/lib/offline/drafts";
 import { formatClock } from "@/lib/scoring/speech-timing";
 import { cn } from "@/lib/utils";
+import { DebateTimer } from "@/components/tournaments/ballot/debate-timer";
+import { ArgumentFlow } from "@/components/tournaments/ballot/argument-flow";
+import { FactCheckingInterface } from "@/components/tournaments/ballot/fact-checking";
+import { CollaborativeNotes } from "@/components/tournaments/ballot/collaborative-notes";
+import { SpeakerPositionManager } from "@/components/tournaments/ballot/speaker-position-manager";
+import { BallotSkeleton } from "@/components/tournaments/ballot/ballot-skeleton";
+import { BallotRow, BallotCard } from "@/components/tournaments/ballot/ballot-list-items";
+import { BallotDetailsDialog } from "@/components/tournaments/ballot/ballot-details-dialog";
+import { FlagBallotDialog } from "@/components/tournaments/ballot/flag-ballot-dialog";
+import { debateStatusColor, debateStatusIcon } from "@/components/tournaments/ballot/debate-status";
+import { SCORING_CATEGORIES } from "@/components/tournaments/ballot/scoring-categories";
+import { useNames } from "@/components/tournaments/ballot/use-names";
 import { useOffline } from "@/hooks/use-offline";
 import { Checkbox } from "@/components/ui/checkbox";
 import { useDebounce } from "@/hooks/use-debounce";
@@ -91,1049 +103,10 @@ interface TournamentBallotsProps {
   schoolId?: Id<"schools">;
 }
 
-const SCORING_CATEGORIES = [
-  {
-    key: "style" as const,
-    label: "Style",
-    icon: Users2,
-    description: "Delivery, clarity, pace, volume, and engagement",
-    color: "text-blue-600",
-    weight: "40%",
-  },
-  {
-    key: "content" as const,
-    label: "Content",
-    icon: Brain,
-    description: "Arguments, evidence, analysis, and rebuttal",
-    color: "text-green-600",
-    weight: "40%",
-  },
-  {
-    key: "strategy" as const,
-    label: "Strategy",
-    icon: Target,
-    description: "Structure, prioritisation, timing, and role fulfilment",
-    color: "text-purple-600",
-    weight: "20%",
-  },
-];
 
-function BallotSkeleton() {
-  return (
-    <div className="space-y-4">
-      <div className="flex flex-col lg:flex-row lg:items-center bg-brown rounded-t-md lg:justify-between gap-4 p-3">
-        <div>
-          <Skeleton className="h-6 w-48 mb-2" />
-          <Skeleton className="h-4 w-64" />
-        </div>
-        <div className="flex items-center gap-3">
-          <Skeleton className="h-8 w-24" />
-          <Skeleton className="h-8 w-18" />
-          <Skeleton className="h-8 w-24" />
-          <Skeleton className="h-8 w-32" />
-          <Skeleton className="h-8 w-8" />
-        </div>
-      </div>
 
-      {Array.from({ length: 6 }).map((_, i) => (
-        <div key={i} className="flex items-center space-x-4 p-4 border rounded-lg">
-          <Skeleton className="h-12 w-12 rounded" />
-          <div className="space-y-2 flex-1">
-            <Skeleton className="h-4 w-[250px]" />
-            <Skeleton className="h-4 w-[200px]" />
-          </div>
-          <div className="flex space-x-2">
-            <Skeleton className="h-8 w-16" />
-            <Skeleton className="h-8 w-16" />
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
 
-function getDebateStatusColor(status: string) {
-  switch (status) {
-    case "completed": return "bg-green-100 text-green-800";
-    case "inProgress": return "bg-blue-100 text-blue-800";
-    case "pending": return "bg-yellow-100 text-yellow-800";
-    case "noShow": return "bg-red-100 text-red-800";
-    default: return "bg-gray-100 text-gray-800";
-  }
-}
 
-function getDebateStatusIcon(status: string) {
-  switch (status) {
-    case "completed": return CheckCircle;
-    case "inProgress": return Timer;
-    case "pending": return Clock;
-    case "noShow": return AlertTriangle;
-    default: return Clock;
-  }
-}
-
-
-interface DebateTimerProps {
-  debate: EnrichedDebate;
-  token: string;
-  tournament?: BallotTournament;
-  position?: SpeakerPosition;
-  onUpdateDebate?: (debate: EnrichedDebate) => void;
-  onTimeUpdate?: (seconds: number) => void;
-  compact?: boolean;
-}
-
-function DebateTimer({ debate, token, tournament, position = "first", onUpdateDebate, compact = false }: DebateTimerProps) {
-  const timer = useSpeechTimer({
-    speakingTimes: tournament?.speaking_times,
-    position,
-  });
-  const { elapsed: currentTime, isRunning } = timer;
-  const [isRecording, setIsRecording] = useState(false);
-  const [recordingDuration, setRecordingDuration] = useState(0);
-  const [mediaRecorder, setMediaRecorder] = useState<MediaRecorder | null>(null);
-  const [, setAudioChunks] = useState<Blob[]>([]);
-  const [isUploading, setIsUploading] = useState(false);
-  const [audioUrl, setAudioUrl] = useState<string | null>(null);
-  const [isPlaying, setIsPlaying] = useState(false);
-
-  const generateUploadUrl = useMutation(api.files.generateUploadUrl);
-  const getFileUrl = useMutation(api.files.getUrl);
-  const updateDebateRecording = useMutation(api.functions.ballots.updateRecording);
-
-  useEffect(() => {
-    let recordingInterval: NodeJS.Timeout;
-    if (isRecording) {
-      recordingInterval = setInterval(() => {
-        setRecordingDuration(prev => prev + 1);
-      }, 1000);
-    }
-    return () => clearInterval(recordingInterval);
-  }, [isRecording]);
-
-  useEffect(() => {
-    const loadRecordingUrl = async () => {
-      if (debate.recording && !audioUrl) {
-        try {
-          const url = await getFileUrl({ storageId: debate.recording });
-          setAudioUrl(url);
-        } catch (error) {
-          console.error('Error loading recording:', error);
-        }
-      }
-    };
-
-    loadRecordingUrl();
-  }, [debate.recording, audioUrl, getFileUrl]);
-
-  const startRecording = async () => {
-    try {
-      const stream = await navigator.mediaDevices.getUserMedia({
-        audio: {
-          echoCancellation: true,
-          noiseSuppression: true,
-          sampleRate: 44100
-        }
-      });
-
-      const recorder = new MediaRecorder(stream, {
-        mimeType: MediaRecorder.isTypeSupported('audio/webm') ? 'audio/webm' : 'audio/mp4'
-      });
-
-      const chunks: Blob[] = [];
-      setAudioChunks(chunks);
-
-      recorder.ondataavailable = (event) => {
-        if (event.data.size > 0) {
-          chunks.push(event.data);
-        }
-      };
-
-      recorder.onstop = async () => {
-        const audioBlob = new Blob(chunks, { type: recorder.mimeType });
-        await uploadRecording(audioBlob);
-        stream.getTracks().forEach(track => track.stop());
-      };
-
-      recorder.start(1000);
-      setMediaRecorder(recorder);
-      setIsRecording(true);
-      setRecordingDuration(0);
-
-      toast.success("Recording started");
-    } catch (error) {
-      console.error('Error starting recording:', error);
-      toast.error("Failed to start recording. Please check microphone permissions.");
-    }
-  };
-
-  const stopRecording = () => {
-    if (mediaRecorder && mediaRecorder.state !== 'inactive') {
-      mediaRecorder.stop();
-      setIsRecording(false);
-      toast.success("Recording stopped. Uploading...");
-    }
-  };
-
-  const uploadRecording = async (audioBlob: Blob) => {
-    setIsUploading(true);
-    try {
-      const uploadUrl = await generateUploadUrl();
-
-      const result = await fetch(uploadUrl, {
-        method: "POST",
-        headers: { "Content-Type": audioBlob.type },
-        body: audioBlob,
-      });
-
-      if (!result.ok) {
-        throw new Error("Upload failed");
-      }
-
-      const { storageId } = await result.json();
-
-      await updateDebateRecording({
-        token,
-        debate_id: debate._id,
-        recording_id: storageId,
-        duration: recordingDuration,
-      });
-
-      const url = await getFileUrl({ storageId });
-      setAudioUrl(url);
-
-      toast.success("Recording uploaded successfully!");
-
-      if (onUpdateDebate) {
-        onUpdateDebate({
-          ...debate,
-          recording: storageId,
-          recording_duration: recordingDuration
-        });
-      }
-    } catch (error) {
-      console.error('Error uploading recording:', error);
-      toast.error("Failed to upload recording");
-    } finally {
-      setIsUploading(false);
-    }
-  };
-
-  const playRecording = () => {
-    if (audioUrl) {
-      const audio = new Audio(audioUrl);
-      audio.play();
-      setIsPlaying(true);
-
-      audio.onended = () => setIsPlaying(false);
-      audio.onerror = () => {
-        setIsPlaying(false);
-        toast.error("Failed to play recording");
-      };
-    }
-  };
-
-  const downloadRecording = () => {
-    if (audioUrl) {
-      const link = document.createElement('a');
-      link.href = audioUrl;
-      link.download = `debate-${debate.room_name}-recording.webm`;
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-    }
-  };
-
-  if (compact) {
-    return (
-      <div className="flex items-center gap-2">
-        <div
-          className={cn(
-            "text-sm font-mono font-bold tabular-nums",
-            timer.isOvertime && "text-destructive",
-            timer.phase === "grace" && "text-amber-600"
-          )}
-        >
-          {timer.display}
-        </div>
-
-        <div className="flex gap-1">
-          <Button
-            onClick={timer.toggle}
-            variant={isRunning ? "destructive" : "default"}
-            size="sm"
-            className="h-5 w-5 p-0"
-          >
-            {isRunning ? <Pause className="h-2 w-2" /> : <Play className="h-2 w-2" />}
-          </Button>
-
-          <Button
-            onClick={timer.reset}
-            variant="outline"
-            size="sm"
-            className="h-5 w-5 p-0"
-          >
-            <Square className="h-2 w-2" />
-          </Button>
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <Card className="p-4">
-      <div className="text-center space-y-4">
-        <div
-          className={cn(
-            "text-2xl md:text-3xl font-bold font-mono tabular-nums",
-            timer.isOvertime && "text-destructive",
-            timer.phase === "grace" && "text-amber-600"
-          )}
-        >
-          {timer.display}
-        </div>
-
-        <div className="space-y-1">
-          <Progress value={timer.progress} className="w-full" />
-          <div className="flex items-center justify-between text-xs text-muted-foreground">
-            <span>{timer.phaseLabel}</span>
-            <span className="tabular-nums">
-              {timer.isOvertime ? "over" : `${timer.remainingDisplay} left`}
-            </span>
-          </div>
-        </div>
-
-        <div className="flex justify-center gap-2 flex-wrap">
-          <Button
-            onClick={timer.toggle}
-            variant={isRunning ? "destructive" : "default"}
-            size="sm"
-          >
-            {isRunning ? <Pause className="h-4 w-4" /> : <Play className="h-4 w-4" />}
-          </Button>
-
-          <Button
-            onClick={timer.reset}
-            variant="outline"
-            size="sm"
-          >
-            <Square className="h-4 w-4" />
-          </Button>
-
-          <Button
-            onClick={isRecording ? stopRecording : startRecording}
-            variant={isRecording ? "destructive" : "outline"}
-            disabled={isUploading}
-            size="sm"
-          >
-            {isUploading ? (
-              <Loader2 className="h-3 w-3 animate-spin" />
-            ) : isRecording ? (
-              <MicOff className="h-4 w-4" />
-            ) : (
-              <Mic className="h-4 w-4" />
-            )}
-          </Button>
-        </div>
-
-        <div className="text-sm text-muted-foreground space-y-1">
-
-          {isRecording && (
-            <div className="text-red-600 font-medium flex items-center justify-center gap-1">
-              <div className="w-2 h-2 bg-red-600 rounded-full animate-pulse"></div>
-              Recording: {formatClock(recordingDuration)}
-            </div>
-          )}
-
-          {debate.recording && audioUrl && (
-            <div className="text-green-600 text-xs space-y-2">
-              <div className="flex items-center justify-center gap-1 text-sm">
-                <CircleCheck className="h-3 w-3" />
-                Recorded ({formatClock(debate.recording_duration || 0)})
-              </div>
-
-              <div className="flex justify-center gap-1">
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  onClick={playRecording}
-                  disabled={isPlaying}
-                  className="h-6 px-2"
-                >
-                  {isPlaying ? (
-                    <Pause className="h-3 w-3" />
-                  ) : (
-                    <Play className="h-3 w-3" />
-                  )}
-                </Button>
-
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  onClick={downloadRecording}
-                  className="h-6 px-2"
-                >
-                  <Download className="h-3 w-3" />
-                </Button>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    </Card>
-  );
-}
-
-interface ArgumentFlowProps {
-  debate: EnrichedDebate;
-  onAddArgument: (argument: ArgumentFlowEntry) => void;
-  onUpdateArgumentFlow: (flow: ArgumentFlowEntry[]) => void;
-}
-
-function ArgumentFlow({ debate, onAddArgument, onUpdateArgumentFlow }: ArgumentFlowProps) {
-  const [newArgument, setNewArgument] = useState("");
-  const [argumentType, setArgumentType] = useState<"main" | "rebuttal" | "poi">("main");
-  const [argumentStrength, setArgumentStrength] = useState(3);
-  const [linkingMode, setLinkingMode] = useState(false);
-  const [selectedArgument, setSelectedArgument] = useState<number | null>(null);
-
-  const argumentFlow = debate.argument_flow || [];
-
-  const handleAddArgument = () => {
-    if (!newArgument.trim()) return;
-
-    const newArg = {
-      type: argumentType,
-      content: newArgument,
-      speaker: debate.current_speaker as Id<"users">,
-      team: (argumentType === "poi"
-        ? debate.opposition_team_id
-        : debate.proposition_team_id) as Id<"teams">,
-      timestamp: Date.now(),
-      strength: argumentStrength,
-      rebutted_by: [],
-    };
-
-    if (!newArg.team) return;
-
-    onAddArgument(newArg);
-    setNewArgument("");
-    setArgumentStrength(3);
-  };
-
-  const handleArgumentConnection = (parentIndex: number, childIndex: number) => {
-    const updatedFlow = argumentFlow.map((arg: any, index: number) => {
-      if (index === parentIndex) {
-        return {
-          ...arg,
-          rebutted_by: [...(arg.rebutted_by || []), childIndex.toString()]
-        };
-      }
-      return arg;
-    });
-    onUpdateArgumentFlow(updatedFlow);
-    setLinkingMode(false);
-    setSelectedArgument(null);
-    toast.success("Arguments linked successfully");
-  };
-
-  const getArgumentColor = (type: string) => {
-    switch (type) {
-      case "main": return "border-blue-500 bg-blue-50";
-      case "rebuttal": return "border-red-500 bg-red-50";
-      case "poi": return "border-yellow-500 bg-yellow-50";
-      default: return "border-gray-500 bg-gray-50";
-    }
-  };
-
-  return (
-    <Card className="p-4">
-      <div className="space-y-4">
-        <div className="flex items-center justify-between">
-          <h4 className="font-medium flex items-center gap-2">
-            <BarChart3 className="h-4 w-4" />
-            Argument Flow
-          </h4>
-          <Button
-            size="sm"
-            variant={linkingMode ? "default" : "outline"}
-            onClick={() => {
-              setLinkingMode(!linkingMode);
-              setSelectedArgument(null);
-            }}
-          >
-            <Link2 className="h-3 w-3 mr-1" />
-            {linkingMode ? "Exit Linking" : "Link Arguments"}
-          </Button>
-        </div>
-
-        {linkingMode && (
-          <Alert>
-            <AlertCircle className="h-4 w-4" />
-            <AlertDescription>
-              Select two arguments to link them. First argument will be marked as rebutted by the second.
-            </AlertDescription>
-          </Alert>
-        )}
-
-        <ScrollArea className="h-64">
-          <div className="space-y-3">
-            {argumentFlow.map((arg: any, index: number) => (
-              <div
-                key={index}
-                className={`p-3 border rounded-lg cursor-pointer transition-all ${getArgumentColor(arg.type)} ${
-                  linkingMode ? 'hover:shadow-md' : ''
-                } ${selectedArgument === index ? 'ring-2 ring-blue-500' : ''}`}
-                onClick={() => {
-                  if (linkingMode) {
-                    if (selectedArgument === null) {
-                      setSelectedArgument(index);
-                    } else if (selectedArgument !== index) {
-                      handleArgumentConnection(selectedArgument, index);
-                    }
-                  }
-                }}
-              >
-                <div className="flex justify-between items-start mb-2">
-                  <Badge variant="outline" className="mb-1">
-                    {arg.type} #{index + 1}
-                  </Badge>
-                  <div className="flex items-center gap-2">
-                    <div className="flex">
-                      {Array.from({ length: 5 }, (_, i) => (
-                        <div
-                          key={i}
-                          className={`w-2 h-2 rounded-full mx-px ${
-                            i < (arg.strength || 3) ? 'bg-green-500' : 'bg-gray-300'
-                          }`}
-                        />
-                      ))}
-                    </div>
-                    <span className="text-xs text-muted-foreground">
-                      {new Date(arg.timestamp).toLocaleTimeString()}
-                    </span>
-                  </div>
-                </div>
-
-                <p className="text-sm mb-2">{arg.content}</p>
-
-                {arg.rebutted_by && arg.rebutted_by.length > 0 && (
-                  <div className="mt-2 pt-2 border-t">
-                    <span className="text-xs text-muted-foreground">Links to:</span>
-                    <div className="flex gap-1 mt-1 flex-wrap">
-                      {arg.rebutted_by.map((rebuttal: string, idx: number) => (
-                        <Badge key={idx} variant="secondary" className="text-xs">
-                          #{parseInt(rebuttal) + 1}
-                        </Badge>
-                      ))}
-                    </div>
-                  </div>
-                )}
-
-                {linkingMode && selectedArgument === index && (
-                  <div className="mt-2 pt-2 border-t">
-                    <span className="text-xs font-medium text-blue-600">Selected - Choose target argument</span>
-                  </div>
-                )}
-              </div>
-            ))}
-          </div>
-        </ScrollArea>
-
-        <div className="space-y-3 pt-3 border-t">
-          <div className="flex gap-2">
-            <Select value={argumentType} onValueChange={(value: any) => setArgumentType(value)}>
-              <SelectTrigger className="w-24">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="main">Main</SelectItem>
-                <SelectItem value="rebuttal">Rebuttal</SelectItem>
-                <SelectItem value="poi">POI</SelectItem>
-              </SelectContent>
-            </Select>
-
-            <div className="flex-1">
-              <Input
-                placeholder="Add argument..."
-                value={newArgument}
-                onChange={(e) => setNewArgument(e.target.value)}
-                onKeyDown={(e) => e.key === 'Enter' && handleAddArgument()}
-              />
-            </div>
-          </div>
-
-          <div className="flex items-center gap-2">
-            <Label className="text-xs">Strength:</Label>
-            <div className="flex items-center gap-1">
-              <Button
-                size="sm"
-                variant="ghost"
-                onClick={() => setArgumentStrength(Math.max(1, argumentStrength - 1))}
-                className="h-6 w-6 p-0"
-              >
-                <Minus className="h-3 w-3" />
-              </Button>
-
-              <div className="flex mx-2">
-                {Array.from({ length: 5 }, (_, i) => (
-                  <div
-                    key={i}
-                    className={`w-3 h-3 rounded-full mx-px cursor-pointer ${
-                      i < argumentStrength ? 'bg-green-500' : 'bg-gray-300'
-                    }`}
-                    onClick={() => setArgumentStrength(i + 1)}
-                  />
-                ))}
-              </div>
-
-              <Button
-                size="sm"
-                variant="ghost"
-                onClick={() => setArgumentStrength(Math.min(5, argumentStrength + 1))}
-                className="h-6 w-6 p-0"
-              >
-                <Plus className="h-3 w-3" />
-              </Button>
-            </div>
-          </div>
-
-          <Button onClick={handleAddArgument} size="sm" className="w-full">
-            Add Argument
-          </Button>
-        </div>
-      </div>
-    </Card>
-  );
-}
-
-interface FactCheckingInterfaceProps {
-  debate: EnrichedDebate;
-  onAddFactCheck: (factCheck: FactCheckEntry) => void;
-  userId: Id<"users">;
-}
-
-function FactCheckingInterface({ debate, onAddFactCheck, userId }: FactCheckingInterfaceProps) {
-  const { factCheckClaim, isFactChecking } = useGemini();
-  const [selectedText, setSelectedText] = useState("");
-  const [factCheckResult, setFactCheckResult] = useState<any>(null);
-  const [manualResult, setManualResult] = useState<"true" | "false" | "partially_true" | "inconclusive">("inconclusive");
-  const [manualExplanation, setManualExplanation] = useState("");
-
-  const existingFactChecks = debate.fact_checks || [];
-
-  const handleAIFactCheck = async () => {
-    if (!selectedText.trim()) return;
-
-    const result = await factCheckClaim(selectedText);
-    setFactCheckResult(result);
-  };
-
-  const handleManualFactCheck = () => {
-    if (!selectedText.trim() || !manualExplanation.trim()) return;
-
-    const factCheck = {
-      claim: selectedText,
-      result: manualResult,
-      explanation: manualExplanation,
-      sources: [],
-      checked_by: userId,
-      timestamp: Date.now(),
-    };
-
-    onAddFactCheck(factCheck);
-    setSelectedText("");
-    setManualExplanation("");
-    setFactCheckResult(null);
-  };
-
-  const handleAcceptAIResult = () => {
-    if (!factCheckResult) return;
-
-    const factCheck = {
-      claim: selectedText,
-      result: factCheckResult.result,
-      explanation: factCheckResult.explanation,
-      sources: factCheckResult.sources || [],
-      checked_by: userId,
-      timestamp: Date.now(),
-    };
-
-    onAddFactCheck(factCheck);
-    setSelectedText("");
-    setFactCheckResult(null);
-  };
-
-  const getResultColor = (result: string) => {
-    switch (result) {
-      case "true": return "text-green-600 bg-green-50 border-green-200";
-      case "false": return "text-red-600 bg-red-50 border-red-200";
-      case "partially_true": return "text-yellow-600 bg-yellow-50 border-yellow-200";
-      case "inconclusive": return "text-gray-600 bg-gray-50 border-gray-200";
-      default: return "text-gray-600 bg-gray-50 border-gray-200";
-    }
-  };
-
-  return (
-    <Card className="p-4">
-      <div className="space-y-4">
-        <h4 className="font-medium flex items-center gap-2">
-          <Search className="h-4 w-4" />
-          Fact Checking
-        </h4>
-
-        <div className="space-y-3">
-          <div>
-            <Label className="text-sm">Claim to check:</Label>
-            <Textarea
-              placeholder="Enter or paste claim to fact-check..."
-              value={selectedText}
-              onChange={(e) => setSelectedText(e.target.value)}
-              rows={2}
-            />
-          </div>
-
-          <div className="flex gap-2 flex-wrap">
-            <Button
-              onClick={handleAIFactCheck}
-              disabled={!selectedText.trim() || isFactChecking}
-              size="sm"
-              variant="outline"
-            >
-              {isFactChecking ? (
-                <>
-                  <Loader2 className="h-3 w-3 animate-spin mr-1" />
-                  Checking...
-                </>
-              ) : (
-                <>
-                  <Brain className="h-3 w-3 mr-1" />
-                  AI Check
-                </>
-              )}
-            </Button>
-
-            <Button
-              onClick={() => setFactCheckResult({ isManual: true })}
-              disabled={!selectedText.trim()}
-              size="sm"
-              variant="outline"
-            >
-              <CheckSquare className="h-3 w-3 mr-1" />
-              Manual Check
-            </Button>
-          </div>
-
-          {factCheckResult && (
-            <div className="space-y-3 p-3 border rounded-lg">
-              {factCheckResult.isManual ? (
-                <div className="space-y-3">
-                  <h5 className="font-medium">Manual Fact Check</h5>
-
-                  <div>
-                    <Label className="text-sm">Result:</Label>
-                    <Select value={manualResult} onValueChange={(value: any) => setManualResult(value)}>
-                      <SelectTrigger>
-                        <SelectValue />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="true">True</SelectItem>
-                        <SelectItem value="false">False</SelectItem>
-                        <SelectItem value="partially_true">Partially True</SelectItem>
-                        <SelectItem value="inconclusive">Inconclusive</SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-
-                  <div>
-                    <Label className="text-sm">Explanation:</Label>
-                    <Textarea
-                      placeholder="Explain your fact check..."
-                      value={manualExplanation}
-                      onChange={(e) => setManualExplanation(e.target.value)}
-                      rows={3}
-                    />
-                  </div>
-
-                  <div className="flex gap-2">
-                    <Button onClick={handleManualFactCheck} size="sm">
-                      Save Fact Check
-                    </Button>
-                    <Button
-                      onClick={() => setFactCheckResult(null)}
-                      size="sm"
-                      variant="ghost"
-                    >
-                      Cancel
-                    </Button>
-                  </div>
-                </div>
-              ) : (
-                <div className="space-y-3">
-                  <h5 className="font-medium">AI Fact Check Result</h5>
-
-                  <div className={`p-2 rounded border ${getResultColor(factCheckResult.result)}`}>
-                    <div className="font-medium capitalize">{factCheckResult.result.replace('_', ' ')}</div>
-                    <div className="text-sm mt-1">
-                      Confidence: {(factCheckResult.confidence * 100).toFixed(0)}%
-                    </div>
-                  </div>
-
-                  {factCheckResult.explanation && (
-                    <div>
-                      <Label className="text-sm">Explanation:</Label>
-                      <p className="text-sm p-2 bg-muted rounded">{factCheckResult.explanation}</p>
-                    </div>
-                  )}
-
-                  {factCheckResult.sources && factCheckResult.sources.length > 0 && (
-                    <div>
-                      <Label className="text-sm">Sources:</Label>
-                      <ul className="text-sm list-disc list-inside p-2 bg-muted rounded">
-                        {factCheckResult.sources.map((source: string, idx: number) => (
-                          <li key={idx}>{source}</li>
-                        ))}
-                      </ul>
-                    </div>
-                  )}
-
-                  <div className="flex gap-2">
-                    <Button onClick={handleAcceptAIResult} size="sm">
-                      Accept & Save
-                    </Button>
-                    <Button
-                      onClick={() => setFactCheckResult({ isManual: true })}
-                      size="sm"
-                      variant="outline"
-                    >
-                      Edit Manually
-                    </Button>
-                    <Button
-                      onClick={() => setFactCheckResult(null)}
-                      size="sm"
-                      variant="ghost"
-                    >
-                      Cancel
-                    </Button>
-                  </div>
-                </div>
-              )}
-            </div>
-          )}
-        </div>
-
-        {existingFactChecks.length > 0 && (
-          <div className="space-y-2">
-            <Label className="text-sm font-medium">Previous Fact Checks:</Label>
-            <ScrollArea className="h-32">
-              {existingFactChecks.map((check: any, idx: number) => (
-                <div key={idx} className={`p-2 mb-2 rounded border text-sm ${getResultColor(check.result)}`}>
-                  <div className="font-medium">&#34;{check.claim}&#34;</div>
-                  <div className="capitalize">{check.result.replace('_', ' ')}</div>
-                  {check.explanation && <div className="text-xs mt-1">{check.explanation}</div>}
-                </div>
-              ))}
-            </ScrollArea>
-          </div>
-        )}
-      </div>
-    </Card>
-  );
-}
-
-interface CollaborativeNotesProps {
-  debate: EnrichedDebate;
-  userId: Id<"users">;
-  onUpdateNotes: (note: SharedNoteEntry) => void;
-  token: string;
-}
-
-function CollaborativeNotes({ debate, userId, onUpdateNotes, token }: CollaborativeNotesProps) {
-  const [notes, setNotes] = useState("");
-  const [visibility, setVisibility] = useState<"private" | "judges" | "all">("judges");
-  const userNames = useNames(token, [userId]);
-
-  const existingNotes = debate.shared_notes || [];
-
-  const handleSaveNote = () => {
-    if (!notes.trim()) return;
-    const nameEntry = userNames?.find((entry) => entry.id === userId);
-    const fullName = nameEntry?.name ?? "Judge";
-    const userName = fullName.trim().split(" ")[0];
-
-    const newNote = {
-      content: notes,
-      author: userId,
-      name: userName,
-      timestamp: Date.now(),
-      visibility,
-    };
-
-    onUpdateNotes(newNote);
-    setNotes("");
-    toast.success("Note added successfully");
-  };
-
-  return (
-    <Card className="p-4">
-      <div className="space-y-4">
-        <div className="flex items-center justify-between">
-          <h4 className="font-medium flex items-center gap-2">
-            <MessageSquare className="h-4 w-4" />
-            Collaborative Notes
-          </h4>
-          <Select value={visibility} onValueChange={(value: any) => setVisibility(value)}>
-            <SelectTrigger className="w-20 text-sm">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="private">Private</SelectItem>
-              <SelectItem value="judges">Judges</SelectItem>
-              <SelectItem value="all">All</SelectItem>
-            </SelectContent>
-          </Select>
-        </div>
-
-        <ScrollArea className="h-32">
-          <div className="space-y-2">
-            {existingNotes.map((note: any, index: number) => (
-              <div key={index} className="p-2 bg-muted rounded text-sm">
-                <div className="flex justify-between items-start mb-1">
-                  <span className="font-medium text-sm">Judge {note.name}</span>
-                  <div className="flex items-center gap-2">
-                    <Badge variant="outline" className="text-xs">
-                      {note.visibility}
-                    </Badge>
-                    <span className="text-xs text-muted-foreground">
-                      {new Date(note.timestamp).toLocaleTimeString()}
-                    </span>
-                  </div>
-                </div>
-                <p className="text-xs">{note.content}</p>
-              </div>
-            ))}
-            {existingNotes.length === 0 && (
-              <div className="text-center text-muted-foreground text-xs py-4">
-                No notes yet. Add the first note below.
-              </div>
-            )}
-          </div>
-        </ScrollArea>
-
-        <div className="space-y-2">
-          <div className="flex gap-2">
-            <Textarea
-              placeholder="Add a note..."
-              value={notes}
-              onChange={(e) => setNotes(e.target.value)}
-              rows={2}
-              className="flex-1"
-            />
-            <Button
-              onClick={handleSaveNote}
-              size="sm"
-              disabled={!notes.trim()}
-              className="self-end"
-            >
-              <Send className="h-4 w-4" />
-            </Button>
-          </div>
-        </div>
-      </div>
-    </Card>
-  );
-}
-
-interface SpeakerPositionManagerProps {
-  speakers: Array<{ id: string; name: string }>;
-  positions: PositionsBySpeaker;
-  onUpdatePositions: (positions: PositionsBySpeaker) => void;
-  tournament: BallotTournament;
-  debate: EnrichedDebate;
-  teamId: string;
-  teamName: string;
-}
-
-function SpeakerPositionManager({ speakers, positions, onUpdatePositions, tournament, debate, teamId, teamName }: SpeakerPositionManagerProps) {
-  const [speakerPositions, setSpeakerPositions] = useState<PositionsBySpeaker>(positions || {});
-
-  const availablePositions = useMemo(() => {
-
-    const wsdcPositions: Array<{ id: SpeakerPosition; label: string }> = [
-      { id: "first", label: "1st Speaker" },
-      { id: "second", label: "2nd Speaker" },
-      { id: "third", label: "3rd Speaker" },
-      { id: "reply", label: "Reply Speaker" },
-    ];
-
-    return wsdcPositions.slice(0, Math.max(speakers.length, tournament?.team_size ?? 3));
-  }, [speakers.length, tournament?.team_size]);
-
-  const handlePositionChange = (speakerId: string, newPosition: SpeakerPosition) => {
-    const newPositions: PositionsBySpeaker = { ...speakerPositions };
-
-    const currentSpeakerWithPosition = Object.keys(newPositions).find(
-      id => newPositions[id] === newPosition && speakers.some((s) => s.id === id)
-    );
-
-    if (currentSpeakerWithPosition) {
-      const displaced = speakerPositions[speakerId];
-      if (displaced) {
-        newPositions[currentSpeakerWithPosition] = displaced;
-      } else {
-        delete newPositions[currentSpeakerWithPosition];
-      }
-    }
-
-    newPositions[speakerId] = newPosition;
-    setSpeakerPositions(newPositions);
-    onUpdatePositions(newPositions);
-  };
-
-  return (
-    <Card className="p-3">
-      <div className="space-y-3">
-        <h4 className="font-medium text-sm">{teamName} - Speaker Positions</h4>
-
-        <div className="space-y-2">
-          {speakers.map((speaker: any) => (
-            <div key={speaker.id} className="flex items-center gap-2 p-2 border rounded text-sm">
-              <div className="flex-1 min-w-0">
-                <p className="font-medium truncate">{speaker.name}</p>
-                <p className="text-xs text-muted-foreground">
-                  {availablePositions.find(p => p.id === speakerPositions[speaker.id])?.label || "Unassigned"}
-                </p>
-              </div>
-
-              <Select
-                value={speakerPositions[speaker.id] || ""}
-                onValueChange={(value) => handlePositionChange(speaker.id, value as SpeakerPosition)}
-              >
-                <SelectTrigger className="w-28">
-                  <SelectValue placeholder="Position" />
-                </SelectTrigger>
-                <SelectContent>
-                  {availablePositions.map((position) => (
-                    <SelectItem key={position.id} value={position.id}>
-                      {position.label}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-          ))}
-        </div>
-      </div>
-    </Card>
-  );
-}
-
-export function useNames(token: string, userIds: string[]) {
-  return useOffline(useQuery(api.functions.ballots.getUserNames, {
-    token,
-    user_ids: userIds as Id<"users">[],
-  }), "speaker names");
-}
 
 interface JudgingInterfaceProps {
   debate: EnrichedDebate;
@@ -2395,1312 +1368,6 @@ function JudgingInterface({ debate, ballot, userId, onSubmitBallot, tournament, 
         </CardContent>
       </Card>
     </div>
-  );
-}
-
-interface BallotListItemProps {
-  debate: EnrichedDebate;
-  userRole: UserRole;
-  userId?: Id<"users">;
-  onViewDetails: (debate: EnrichedDebate) => void;
-  onEditBallot: (debate: EnrichedDebate) => void;
-  onFlagBallot: (debate: EnrichedDebate) => void;
-  onUnflagBallot: (debate: EnrichedDebate) => void;
-}
-
-function BallotRow({ debate, userRole, userId, onViewDetails, onEditBallot, onFlagBallot, onUnflagBallot }: BallotListItemProps) {
-  const StatusIcon = getDebateStatusIcon(debate.status);
-  const canEdit = userRole === "admin" || (userRole === "volunteer" && debate.judges?.some((j: any) => j._id === userId));
-  const canSeeDetails = debate.can_see_full_details || userRole === "admin" || userRole === "volunteer";
-  const canFlag = userRole === "admin" || (userRole === "volunteer" && debate.judges?.some((j: any) => j._id === userId));
-  const hasFlaggedBallots = debate.has_flagged_ballots || debate.judges?.some((j: any) => j.is_flagged);
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const checkIsMobile = () => setIsMobile(window.innerWidth < 768);
-    checkIsMobile();
-    window.addEventListener('resize', checkIsMobile);
-    return () => window.removeEventListener('resize', checkIsMobile);
-  }, []);
-
-  const handleEditClick = () => {
-    if (isMobile) {
-      onEditBallot(debate);
-    } else {
-      onEditBallot(debate);
-    }
-  };
-
-  return (
-    <TableRow className="hover:bg-muted/50">
-      <TableCell>
-        <div className="space-y-1">
-          <div className="flex items-center gap-2">
-            <span className="font-medium">{debate.room_name}</span>
-            <Badge variant="secondary" className={getDebateStatusColor(debate.status)}>
-              <StatusIcon className="h-3 w-3 mr-1" />
-              {debate.status}
-            </Badge>
-            {hasFlaggedBallots && (
-              <Badge variant="destructive" className="gap-1">
-                <Flag className="h-3 w-3" />
-                Flagged
-              </Badge>
-            )}
-          </div>
-          <div className="text-sm text-muted-foreground">
-            Round {debate.round?.round_number} • {debate.round?.type}
-          </div>
-        </div>
-      </TableCell>
-
-      <TableCell>
-        <div className="space-y-2">
-          {debate.proposition_team && (
-            <div className="flex items-center gap-2">
-              <Badge variant="outline" className="text-green-700 border-green-700 text-xs">Prop</Badge>
-              <span className="text-sm">{debate.proposition_team.name}</span>
-              {debate.winning_team_id === debate.proposition_team._id && (
-                <Crown className="h-3 w-3 text-yellow-500" />
-              )}
-            </div>
-          )}
-          {debate.opposition_team && (
-            <div className="flex items-center gap-2">
-              <Badge variant="outline" className="text-red-700 border-red-700 text-xs">Opp</Badge>
-              <span className="text-sm">{debate.opposition_team.name}</span>
-              {debate.winning_team_id === debate.opposition_team._id && (
-                <Crown className="h-3 w-3 text-yellow-500" />
-              )}
-            </div>
-          )}
-        </div>
-      </TableCell>
-
-      <TableCell>
-        <div className="flex flex-wrap gap-1">
-          {debate.judges?.slice(0, 3).map((judge: any) => (
-            <Badge
-              key={judge._id}
-              variant={judge.is_head_judge ? "default" : "outline"}
-              className={`text-xs ${judge.is_flagged ? 'border-red-500 text-red-600' : ''}`}
-            >
-              {judge.name?.split(' ')[0] || `J${judge._id.slice(-3)}`}
-              {judge.is_head_judge && <Crown className="h-2 w-2 ml-1" />}
-              {userRole === "admin" && (
-                <div className="ml-1">
-                  {judge.is_final ? (
-                    <CheckCircle className="h-2 w-2 text-green-500" />
-                  ) : judge.has_submitted ? (
-                    <Clock className="h-2 w-2 text-yellow-500" />
-                  ) : (
-                    <div className="h-2 w-2 rounded-full bg-gray-300" />
-                  )}
-                </div>
-              )}
-            </Badge>
-          ))}
-          {debate.judges?.length > 3 && (
-            <Badge variant="outline" className="text-xs">
-              +{debate.judges.length - 3}
-            </Badge>
-          )}
-        </div>
-      </TableCell>
-
-      <TableCell>
-        {userRole === "admin" && debate.judges?.length > 0 && (
-          <div className="space-y-1">
-            <div className="text-sm">
-              {debate.final_submissions_count || 0}/{debate.judges.length}
-            </div>
-            <Progress
-              value={debate.judges.length > 0 ? (debate.final_submissions_count || 0) / debate.judges.length * 100 : 0}
-              className="w-16 h-1"
-            />
-          </div>
-        )}
-      </TableCell>
-
-      <TableCell>
-        <div className="flex items-center gap-1">
-          {debate.recording && (
-            <Button variant="ghost" size="sm" title="Download Recording" className="h-6 w-6 p-0">
-              <Download className="h-3 w-3" />
-            </Button>
-          )}
-          {canFlag && debate.status !== "pending" && (
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => onFlagBallot(debate)}
-              title="Flag Ballot"
-              className={`h-6 w-6 p-0 ${hasFlaggedBallots ? "text-red-600" : ""}`}
-            >
-              <Flag className="h-3 w-3" />
-            </Button>
-          )}
-          {userRole === "admin" && hasFlaggedBallots && (
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => onUnflagBallot(debate)}
-              title="Unflag Ballot"
-              className="h-6 w-6 p-0 text-green-600"
-            >
-              <CheckCircle className="h-3 w-3" />
-            </Button>
-          )}
-          {canEdit && (
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={handleEditClick}
-              title="Edit Ballot"
-              className="h-6 w-6 p-0"
-            >
-              <Edit3 className="h-3 w-3" />
-            </Button>
-          )}
-          {canSeeDetails && (
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => onViewDetails(debate)}
-              title="View Details"
-              className="h-6 w-6 p-0"
-            >
-              <Eye className="h-3 w-3" />
-            </Button>
-          )}
-        </div>
-      </TableCell>
-    </TableRow>
-  );
-}
-
-function BallotCard({ debate, userRole, userId, onViewDetails, onEditBallot, onFlagBallot, onUnflagBallot }: BallotListItemProps) {
-  const StatusIcon = getDebateStatusIcon(debate.status);
-  const canEdit = userRole === "admin" || (userRole === "volunteer" && debate.judges?.some((j: any) => j._id === userId));
-  const canSeeDetails = debate.can_see_full_details || userRole === "admin" || userRole === "volunteer";
-  const canFlag = userRole === "admin" || (userRole === "volunteer" && debate.judges?.some((j: any) => j._id === userId));
-  const hasFlaggedBallots = debate.has_flagged_ballots || debate.judges?.some((j: any) => j.is_flagged);
-  const submissionProgress = debate.judges?.length > 0 ? (debate.final_submissions_count || 0) / debate.judges.length * 100 : 0;
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const checkIsMobile = () => setIsMobile(window.innerWidth < 768);
-    checkIsMobile();
-    window.addEventListener('resize', checkIsMobile);
-    return () => window.removeEventListener('resize', checkIsMobile);
-  }, []);
-
-  const handleEditClick = () => {
-    if (isMobile) {
-      onEditBallot(debate);
-    } else {
-      onEditBallot(debate);
-    }
-  };
-
-  return (
-    <Card className="hover:shadow-md transition-all duration-200">
-      <CardHeader className="pb-3">
-        <div className="flex items-center justify-between">
-          <div className="flex items-center gap-3">
-            <CardTitle className="text-lg">{debate.room_name}</CardTitle>
-            <Badge variant="secondary" className={getDebateStatusColor(debate.status)}>
-              <StatusIcon className="h-3 w-3 mr-1" />
-              {debate.status}
-            </Badge>
-            {hasFlaggedBallots && (
-              <Badge variant="destructive" className="gap-1">
-                <Flag className="h-3 w-3" />
-                Flagged
-              </Badge>
-            )}
-          </div>
-          <div className="flex items-center gap-2">
-            {debate.recording && (
-              <Button variant="ghost" size="sm" title="Download Recording">
-                <Download className="h-4 w-4" />
-              </Button>
-            )}
-            {canFlag && debate.status !== "pending" && (
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => onFlagBallot(debate)}
-                title="Flag Ballot"
-                className={hasFlaggedBallots ? "text-red-600" : ""}
-              >
-                <Flag className="h-4 w-4" />
-              </Button>
-            )}
-            {userRole === "admin" && hasFlaggedBallots && (
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={() => onUnflagBallot(debate)}
-                title="Unflag Ballot"
-                className="text-green-600"
-              >
-                <CheckCircle className="h-4 w-4" />
-              </Button>
-            )}
-            {canEdit && (
-              <Button variant="ghost" size="sm" onClick={handleEditClick} title="Edit Ballot">
-                <Edit3 className="h-4 w-4" />
-              </Button>
-            )}
-            {canSeeDetails && (
-              <Button variant="ghost" size="sm" onClick={() => onViewDetails(debate)} title="View Details">
-                <Eye className="h-4 w-4" />
-              </Button>
-            )}
-          </div>
-        </div>
-      </CardHeader>
-
-      <CardContent className="space-y-4">
-
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
-          {debate.proposition_team && (
-            <div className="flex items-center gap-2">
-              {debate.winning_team_id === debate.proposition_team._id && (
-                <Crown className="h-4 w-4 text-yellow-500" />
-              )}
-              <Badge variant="outline" className="text-green-700 border-green-700">Prop</Badge>
-              <div className="min-w-0 flex-1">
-                <p className="font-medium truncate">{debate.proposition_team.name}</p>
-                {debate.proposition_team.school && (
-                  <p className="text-xs text-muted-foreground truncate">
-                    {debate.proposition_team.school.name}
-                  </p>
-                )}
-              </div>
-            </div>
-          )}
-
-          {debate.opposition_team && (
-            <div className="flex items-center gap-2">
-              <Badge variant="outline" className="text-red-700 border-red-700">Opp</Badge>
-              <div className="min-w-0 flex-1">
-                <p className="font-medium truncate">{debate.opposition_team.name}</p>
-                {debate.opposition_team.school && (
-                  <p className="text-xs text-muted-foreground truncate">
-                    {debate.opposition_team.school.name}
-                  </p>
-                )}
-              </div>
-              {debate.winning_team_id === debate.opposition_team._id && (
-                <Crown className="h-4 w-4 text-yellow-500" />
-              )}
-            </div>
-          )}
-        </div>
-
-        <Separator />
-
-
-        <div>
-          <div className="flex items-center justify-between mb-2">
-            <span className="text-sm font-medium">
-              Judges ({debate.judges?.length || 0})
-            </span>
-            {userRole === "admin" && (
-              <div className="text-xs text-muted-foreground">
-                {submissionProgress.toFixed(0)}% submitted
-              </div>
-            )}
-          </div>
-
-          <div className="flex flex-wrap gap-2">
-            {debate.judges?.map((judge: any) => (
-              <Badge
-                key={judge._id}
-                variant={judge.is_head_judge ? "default" : "outline"}
-                className={`gap-1 ${judge.is_flagged ? 'border-red-500 text-red-600' : ''}`}
-              >
-                {judge.name?.split(' ')[0] || `Judge ${judge._id.slice(-3)}`}
-                {judge.is_head_judge && <Crown className="h-3 w-3" />}
-                {judge.is_flagged && <Flag className="h-3 w-3" />}
-                {userRole === "admin" && (
-                  <div className="ml-1">
-                    {judge.is_final ? (
-                      <CheckCircle className="h-3 w-3 text-green-500" />
-                    ) : judge.has_submitted ? (
-                      <Clock className="h-3 w-3 text-yellow-500" />
-                    ) : (
-                      <div className="h-3 w-3 rounded-full bg-gray-300" />
-                    )}
-                  </div>
-                )}
-              </Badge>
-            ))}
-          </div>
-        </div>
-
-
-        <div className="flex items-center justify-between text-sm text-muted-foreground">
-          <span>Round {debate.round?.round_number}</span>
-          <span>{debate.round?.type}</span>
-        </div>
-
-
-        {userRole === "admin" && debate.judges?.length > 0 && (
-          <div className="space-y-1">
-            <div className="flex justify-between text-xs">
-              <span>Submission Progress</span>
-              <span>{debate.final_submissions_count || 0}/{debate.judges.length}</span>
-            </div>
-            <Progress value={submissionProgress} className="w-full h-2" />
-          </div>
-        )}
-
-
-        {debate.winning_team_id && canSeeDetails && (
-          <div className="pt-2 border-t">
-            <div className="flex items-center justify-between text-sm">
-              <span className="text-muted-foreground">Result:</span>
-              <div className="flex items-center gap-2">
-                <span className="font-medium">
-                  {debate.winning_team_id === debate.proposition_team?._id
-                    ? debate.proposition_team.name
-                    : debate.opposition_team?.name
-                  } wins
-                </span>
-              </div>
-            </div>
-          </div>
-        )}
-
-
-        {((debate.fact_checks?.length ?? 0) > 0 || (debate.argument_flow?.length ?? 0) > 0) && canSeeDetails && (
-          <div className="pt-2 border-t">
-            <div className="flex items-center gap-4 text-xs text-muted-foreground">
-              {(debate.fact_checks?.length ?? 0) > 0 && (
-                <span className="flex items-center gap-1">
-                  <Search className="h-3 w-3" />
-                  {(debate.fact_checks?.length ?? 0)} fact checks
-                </span>
-              )}
-              {(debate.argument_flow?.length ?? 0) > 0 && (
-                <span className="flex items-center gap-1">
-                  <BarChart3 className="h-3 w-3" />
-                  {(debate.argument_flow?.length ?? 0)} arguments
-                </span>
-              )}
-            </div>
-          </div>
-        )}
-      </CardContent>
-    </Card>
-  );
-}
-
-interface BallotDetailsDialogProps {
-  debate: EnrichedDebate | null;
-  isOpen: boolean;
-  onClose: () => void;
-  token: string;
-  userRole?: UserRole;
-}
-
-function BallotDetailsDialog({ debate, isOpen, onClose, token }: BallotDetailsDialogProps) {
-  const [selectedJudge, setSelectedJudge] = useState<string>("all");
-  const [expandedJudge, setExpandedJudge] = useState<string | null>(null);
-  const [expandedSpeaker, setExpandedSpeaker] = useState<string | null>(null);
-
-  const allUserIds = useMemo(() => {
-    const userIds = new Set<string>();
-
-    if (debate?.judges) {
-      debate.judges.forEach((judge: any) => {
-        const judgeId = judge._id || judge;
-        userIds.add(judgeId);
-      });
-    }
-
-    if (debate?.judges_ballots) {
-      debate.judges_ballots.forEach((judgeBallot: any) => {
-        if (judgeBallot.ballot?.speaker_scores) {
-          judgeBallot.ballot.speaker_scores.forEach((score: any) => {
-            userIds.add(score.speaker_id);
-          });
-        }
-      });
-    }
-
-    if (debate?.proposition_team?.members) {
-      debate.proposition_team.members.forEach((memberId: string) => {
-        userIds.add(memberId);
-      });
-    }
-    if (debate?.opposition_team?.members) {
-      debate.opposition_team.members.forEach((memberId: string) => {
-        userIds.add(memberId);
-      });
-    }
-
-    return Array.from(userIds);
-  }, [debate]);
-
-  const userNamesQuery = useNames(token, allUserIds);
-
-  const getUserName = (userId: string) => {
-    const user = userNamesQuery?.find((u: any) => u.id === userId);
-    return user?.name || `User ${userId.slice(-4)}`;
-  };
-
-  const getJudgeName = (judgeId: string) => {
-    const judge = debate?.judges?.find(
-      (j) => (typeof j === "string" ? j : j._id) === judgeId
-    );
-
-    if (judge && typeof judge !== "string" && judge.name) return judge.name;
-
-    return getUserName(judgeId);
-  };
-
-  const ballotDetails = useMemo(() => {
-    if (!debate?.judges_ballots) return [];
-
-    return debate.judges_ballots.map((judgeBallot: any) => {
-      if (!judgeBallot.ballot) return null;
-
-      return {
-        ...judgeBallot.ballot,
-        judge_id: judgeBallot.judge_id,
-        judge_name: getJudgeName(judgeBallot.judge_id),
-        is_head_judge: debate.head_judge_id === judgeBallot.judge_id,
-        is_flagged: judgeBallot.ballot.flagged ?? false,
-      };
-    }).filter(Boolean);
-  }, [debate, userNamesQuery]);
-
-  if (!debate) return null;
-
-  const filteredBallots = selectedJudge === "all"
-    ? ballotDetails
-    : ballotDetails.filter((b: any) => b.judge_id === selectedJudge);
-
-  const getWinningTeamName = () => {
-    if (debate?.winning_team_id === debate?.proposition_team?._id) {
-      return debate.proposition_team?.name;
-    } else if (debate?.winning_team_id === debate?.opposition_team?._id) {
-      return debate.opposition_team?.name;
-    }
-    return "Unknown";
-  };
-
-  const getWinningPosition = () => {
-    if (debate?.winning_team_id === debate?.proposition_team?._id) {
-      return "proposition";
-    } else if (debate?.winning_team_id === debate?.opposition_team?._id) {
-      return "opposition";
-    }
-    return "unknown";
-  };
-
-  const groupSpeakerScoresByTeam = (speakerScores: any[]) => {
-    const propTeamScores = speakerScores.filter(score =>
-      score.team_id === debate?.proposition_team?._id
-    );
-    const oppTeamScores = speakerScores.filter(score =>
-      score.team_id === debate?.opposition_team?._id
-    );
-
-    return {
-      proposition: propTeamScores,
-      opposition: oppTeamScores
-    };
-  };
-
-  const handleJudgeToggle = (judgeId: string) => {
-    if (expandedJudge === judgeId) {
-      setExpandedJudge(null);
-    } else {
-      setExpandedJudge(judgeId);
-      setExpandedSpeaker(null); // Reset speaker expansion when changing judges
-    }
-  };
-
-  const handleSpeakerToggle = (speakerId: string) => {
-    if (expandedSpeaker === speakerId) {
-      setExpandedSpeaker(null);
-    } else {
-      setExpandedSpeaker(speakerId);
-    }
-  };
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent className="max-w-6xl max-h-[95vh] overflow-hidden flex flex-col">
-        <DialogHeader className="flex-shrink-0 border-b pb-4">
-          <DialogTitle className="flex items-center gap-2 text-lg md:text-xl">
-            <FileText className="h-4 w-4 md:h-5 md:w-5" />
-            <span className="truncate">Ballot Details - {debate?.room_name || "Unknown Debate"}</span>
-          </DialogTitle>
-        </DialogHeader>
-
-        <div className="flex-1 overflow-y-auto">
-          <div className="space-y-4 md:space-y-6 p-4">
-            
-            {ballotDetails.length > 1 && (
-              <div className="flex flex-col sm:flex-row sm:items-center gap-2">
-                <Label className="text-sm md:text-base flex-shrink-0">View Judge:</Label>
-                <Select value={selectedJudge} onValueChange={setSelectedJudge}>
-                  <SelectTrigger className="w-full sm:w-48">
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    <SelectItem value="all">All Judges</SelectItem>
-                    {ballotDetails.map((ballot: any) => (
-                      <SelectItem key={ballot.judge_id} value={ballot.judge_id}>
-                        {ballot.judge_name}
-                        {ballot.is_head_judge && " (Head)"}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-              </div>
-            )}
-
-            
-            <div className="grid grid-cols-1 md:grid-cols-2 gap-3 md:gap-4 p-3 md:p-4 bg-muted rounded-lg">
-              <div className="text-center">
-                <h3 className="font-semibold text-green-700 text-sm md:text-base truncate">{debate?.proposition_team?.name}</h3>
-                <p className="text-xs md:text-sm text-muted-foreground">Proposition</p>
-                <div className="text-xs text-muted-foreground mt-1 truncate">
-                  {debate?.proposition_team?.school?.name}
-                </div>
-                {getWinningPosition() === "proposition" && (
-                  <Badge variant="default" className="mt-2 text-xs">
-                    <Crown className="h-3 w-3 mr-1" />
-                    Winner
-                  </Badge>
-                )}
-              </div>
-              <div className="text-center">
-                <h3 className="font-semibold text-red-700 text-sm md:text-base truncate">{debate?.opposition_team?.name}</h3>
-                <p className="text-xs md:text-sm text-muted-foreground">Opposition</p>
-                <div className="text-xs text-muted-foreground mt-1 truncate">
-                  {debate?.opposition_team?.school?.name}
-                </div>
-                {getWinningPosition() === "opposition" && (
-                  <Badge variant="default" className="mt-2 text-xs">
-                    <Crown className="h-3 w-3 mr-1" />
-                    Winner
-                  </Badge>
-                )}
-              </div>
-            </div>
-
-            
-            {(debate?.fact_checks?.length ?? 0) > 0 && (
-              <Card>
-                <CardHeader className="pb-3">
-                  <CardTitle className="text-base md:text-lg flex items-center gap-2">
-                    <Search className="h-4 w-4 md:h-5 md:w-5" />
-                    Fact Checks ({(debate.fact_checks?.length ?? 0)})
-                  </CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <div className="space-y-3">
-                    {(debate.fact_checks ?? []).map((check: any, idx: number) => (
-                      <div key={idx} className={`p-3 rounded border ${
-                        check.result === 'true' ? 'bg-green-50 border-green-200' :
-                          check.result === 'false' ? 'bg-red-50 border-red-200' :
-                            check.result === 'partially_true' ? 'bg-yellow-50 border-yellow-200' :
-                              'bg-gray-50 border-gray-200'
-                      }`}>
-                        <div className="flex flex-col sm:flex-row sm:justify-between sm:items-start gap-2 mb-2">
-                          <Badge variant="outline" className="capitalize text-xs w-fit">
-                            {check.result.replace('_', ' ')}
-                          </Badge>
-                          <span className="text-xs text-muted-foreground">
-                            {new Date(check.timestamp).toLocaleString()}
-                          </span>
-                        </div>
-                        <p className="text-sm font-medium mb-1 break-words">&#34;{check.claim}&#34;</p>
-                        {check.explanation && (
-                          <p className="text-sm text-muted-foreground break-words">{check.explanation}</p>
-                        )}
-                        {check.sources?.length > 0 && (
-                          <div className="mt-2">
-                            <p className="text-xs font-medium">Sources:</p>
-                            <ul className="text-xs list-disc list-inside">
-                              {check.sources.map((source: string, i: number) => (
-                                <li key={i} className="break-words">{source}</li>
-                              ))}
-                            </ul>
-                          </div>
-                        )}
-                      </div>
-                    ))}
-                  </div>
-                </CardContent>
-              </Card>
-            )}
-
-            
-            {(debate?.argument_flow?.length ?? 0) > 0 && (
-              <Card>
-                <CardHeader className="pb-3">
-                  <CardTitle className="text-base md:text-lg flex items-center gap-2">
-                    <BarChart3 className="h-4 w-4 md:h-5 md:w-5" />
-                    Argument Flow ({(debate.argument_flow?.length ?? 0)})
-                  </CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <div className="space-y-3">
-                    {(debate.argument_flow ?? []).map((arg: any, idx: number) => (
-                      <div key={idx} className={`p-3 rounded border ${
-                        arg.type === 'main' ? 'bg-blue-50 border-blue-200' :
-                          arg.type === 'rebuttal' ? 'bg-red-50 border-red-200' :
-                            'bg-yellow-50 border-yellow-200'
-                      }`}>
-                        <div className="flex flex-col sm:flex-row sm:justify-between sm:items-start gap-2 mb-2">
-                          <Badge variant="outline" className="capitalize text-xs w-fit">
-                            {arg.type} #{idx + 1}
-                          </Badge>
-                          <div className="flex flex-col sm:flex-row sm:items-center gap-2">
-                            {arg.strength && (
-                              <div className="flex">
-                                {Array.from({ length: 5 }, (_, i) => (
-                                  <div
-                                    key={i}
-                                    className={`w-2 h-2 rounded-full mx-px ${
-                                      i < arg.strength ? 'bg-sky-500' : 'bg-gray-300'
-                                    }`}
-                                  />
-                                ))}
-                              </div>
-                            )}
-                            <span className="text-xs text-muted-foreground">
-                              {new Date(arg.timestamp).toLocaleString()}
-                            </span>
-                          </div>
-                        </div>
-                        <p className="text-sm break-words">{arg.content}</p>
-                        <p className="text-xs text-muted-foreground mt-1">
-                          Speaker {getUserName(arg.speaker)}
-                        </p>
-                        {arg.rebutted_by?.length > 0 && (
-                          <div className="mt-2 pt-2 border-t">
-                            <span className="text-xs text-muted-foreground">Links to:</span>
-                            <div className="flex gap-1 mt-1 flex-wrap">
-                              {arg.rebutted_by.map((rebuttal: string, i: number) => (
-                                <Badge key={i} variant="secondary" className="text-xs">
-                                  #{parseInt(rebuttal) + 1}
-                                </Badge>
-                              ))}
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    ))}
-                  </div>
-                </CardContent>
-              </Card>
-            )}
-
-            
-            {filteredBallots.map((ballot: any, index: number) => (
-              <Collapsible
-                key={ballot.judge_id || index}
-                open={expandedJudge === ballot.judge_id}
-                onOpenChange={() => handleJudgeToggle(ballot.judge_id)}
-              >
-                <Card>
-                  <CollapsibleTrigger asChild>
-                    <Button
-                      variant="ghost"
-                      className="w-full p-3 md:p-4 h-auto justify-between hover:bg-muted/50"
-                    >
-                      <div className="flex items-center gap-2 min-w-0 flex-1">
-                        {expandedJudge === ballot.judge_id ? (
-                          <ChevronDown className="h-4 w-4 flex-shrink-0" />
-                        ) : (
-                          <ChevronRight className="h-4 w-4 flex-shrink-0" />
-                        )}
-                        <div className="text-left min-w-0 flex-1">
-                          <div className="flex flex-col sm:flex-row sm:items-center gap-2">
-                            <span className="font-medium text-sm md:text-base truncate">{ballot.judge_name}</span>
-                            <div className="flex flex-wrap gap-1">
-                              {ballot.is_head_judge && (
-                                <Badge variant="default" className="text-xs">
-                                  <Crown className="h-3 w-3 mr-1" />
-                                  Head
-                                </Badge>
-                              )}
-                              {ballot.submission_state === "submitted" && (
-                                <Badge variant="outline" className="text-green-600 text-xs">
-                                  <CheckCircle className="h-3 w-3 mr-1" />
-                                  Final
-                                </Badge>
-                              )}
-                              {ballot.is_flagged && (
-                                <Badge variant="destructive" className="text-xs">
-                                  <Flag className="h-3 w-3 mr-1" />
-                                  Flagged
-                                </Badge>
-                              )}
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                      <div className="text-right min-w-0 flex-shrink-0 ml-2">
-                        <div className="text-xs md:text-sm font-medium truncate">
-                          {ballot.winning_team_id === debate?.proposition_team?._id
-                            ? debate?.proposition_team?.name
-                            : ballot.winning_team_id === debate?.opposition_team?._id
-                              ? debate?.opposition_team?.name
-                              : "No decision"
-                          }
-                        </div>
-                      </div>
-                    </Button>
-                  </CollapsibleTrigger>
-
-                  <CollapsibleContent>
-                    <div className="px-3 md:px-4 pb-3 md:pb-4 space-y-4 border-t">
-                      
-                      <div className="p-3 bg-muted rounded-lg mt-4">
-                        <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-2">
-                          <span className="font-medium text-sm md:text-base">Decision:</span>
-                          <div className="flex flex-col sm:flex-row sm:items-center gap-2">
-                            <span className="font-semibold text-sm md:text-base">
-                              {ballot.winning_team_id === debate?.proposition_team?._id
-                                ? debate?.proposition_team?.name
-                                : ballot.winning_team_id === debate?.opposition_team?._id
-                                  ? debate?.opposition_team?.name
-                                  : "No decision recorded"
-                              }
-                            </span>
-                            {ballot.winning_position && (
-                              <Badge variant="outline" className={`text-xs ${
-                                ballot.winning_position === "proposition"
-                                  ? "text-green-700 border-green-700"
-                                  : "text-red-700 border-red-700"
-                              }`}>
-                                {ballot.winning_position}
-                              </Badge>
-                            )}
-                          </div>
-                        </div>
-                      </div>
-
-                      
-                      {ballot.speaker_scores?.length > 0 && (
-                        <div>
-                          <h4 className="font-medium mb-3 text-sm md:text-base">Speaker Scores</h4>
-
-                          
-                          {(() => {
-                            const { proposition, opposition } = groupSpeakerScoresByTeam(ballot.speaker_scores);
-
-                            return (
-                              <div className="space-y-4">
-                                
-                                {proposition.length > 0 && (
-                                  <div>
-                                    <h5 className="font-medium text-green-700 mb-2 flex flex-col sm:flex-row sm:items-center gap-2">
-                                      <Badge variant="outline" className="text-green-700 border-green-700 text-xs w-fit">Prop</Badge>
-                                      <span className="text-sm md:text-base">{debate?.proposition_team?.name}</span>
-                                    </h5>
-                                    <div className="space-y-3">
-                                      {proposition.map((score: any, scoreIndex: number) => (
-                                        <Collapsible
-                                          key={`prop-${score.speaker_id}`}
-                                          open={expandedSpeaker === score.speaker_id}
-                                          onOpenChange={() => handleSpeakerToggle(score.speaker_id)}
-                                        >
-                                          <Card className="border-green-200">
-                                            <CollapsibleTrigger asChild>
-                                              <Button
-                                                variant="ghost"
-                                                className="w-full p-3 md:p-4 h-auto justify-between hover:bg-green-50"
-                                              >
-                                                <div className="flex items-center gap-2 min-w-0 flex-1">
-                                                  {expandedSpeaker === score.speaker_id ? (
-                                                    <ChevronDown className="h-4 w-4 flex-shrink-0" />
-                                                  ) : (
-                                                    <ChevronRight className="h-4 w-4 flex-shrink-0" />
-                                                  )}
-                                                  <div className="text-left min-w-0 flex-1">
-                                                    <h6 className="font-medium text-sm md:text-base truncate">{getUserName(score.speaker_id)}</h6>
-                                                    <p className="text-xs md:text-sm text-muted-foreground">{score.position}</p>
-                                                  </div>
-                                                </div>
-                                                <div className="text-right flex-shrink-0 ml-2">
-                                                  <div className="text-xl md:text-2xl font-bold text-primary">{score.score}</div>
-                                                  <div className="text-xs md:text-sm text-muted-foreground">out of 30</div>
-                                                </div>
-                                              </Button>
-                                            </CollapsibleTrigger>
-
-                                            <CollapsibleContent>
-                                              <div className="px-3 md:px-4 pb-3 md:pb-4 space-y-3 border-t">
-                                                
-                                                <div className="grid grid-cols-2 lg:grid-cols-4 gap-3 mt-4">
-                                                  {SCORING_CATEGORIES.map((category) => {
-                                                    const CategoryIcon = category.icon;
-                                                    const categoryScore = score[category.key] || 0;
-                                                    return (
-                                                      <div key={category.key} className="text-center">
-                                                        <div className="flex items-center justify-center gap-1 mb-1">
-                                                          <CategoryIcon className={`h-3 w-3 ${category.color}`} />
-                                                          <span className="text-xs font-medium truncate">{category.label.split(' ')[0]}</span>
-                                                        </div>
-                                                        <div className="text-sm font-bold tabular-nums">{categoryScore}</div>
-                                                        <Progress value={(categoryScore / 25) * 100} className="h-1 mt-1" />
-                                                      </div>
-                                                    );
-                                                  })}
-                                                </div>
-
-                                                
-                                                {score.comments && (
-                                                  <div className="pt-3 border-t">
-                                                    <Label className="text-xs font-medium">Judge Feedback:</Label>
-                                                    <p className="text-sm mt-1 p-2 bg-muted rounded break-words">{score.comments}</p>
-                                                  </div>
-                                                )}
-
-                                                
-                                                {score.bias_detected && (
-                                                  <div className="pt-3 border-t">
-                                                    <Alert variant="destructive">
-                                                      <AlertCircle className="h-4 w-4" />
-                                                      <AlertDescription>
-                                                        <p className="font-medium">Bias Detected</p>
-                                                        {score.bias_explanation && (
-                                                          <p className="text-sm mt-1 break-words">{score.bias_explanation}</p>
-                                                        )}
-                                                      </AlertDescription>
-                                                    </Alert>
-                                                  </div>
-                                                )}
-                                              </div>
-                                            </CollapsibleContent>
-                                          </Card>
-                                        </Collapsible>
-                                      ))}
-                                    </div>
-                                  </div>
-                                )}
-
-                                
-                                {opposition.length > 0 && (
-                                  <div>
-                                    <h5 className="font-medium text-red-700 mb-2 flex flex-col sm:flex-row sm:items-center gap-2">
-                                      <Badge variant="outline" className="text-red-700 border-red-700 text-xs w-fit">Opp</Badge>
-                                      <span className="text-sm md:text-base">{debate?.opposition_team?.name}</span>
-                                    </h5>
-                                    <div className="space-y-3">
-                                      {opposition.map((score: any, scoreIndex: number) => (
-                                        <Collapsible
-                                          key={`opp-${score.speaker_id}`}
-                                          open={expandedSpeaker === score.speaker_id}
-                                          onOpenChange={() => handleSpeakerToggle(score.speaker_id)}
-                                        >
-                                          <Card className="border-red-200">
-                                            <CollapsibleTrigger asChild>
-                                              <Button
-                                                variant="ghost"
-                                                className="w-full p-3 md:p-4 h-auto justify-between hover:bg-red-50"
-                                              >
-                                                <div className="flex items-center gap-2 min-w-0 flex-1">
-                                                  {expandedSpeaker === score.speaker_id ? (
-                                                    <ChevronDown className="h-4 w-4 flex-shrink-0" />
-                                                  ) : (
-                                                    <ChevronRight className="h-4 w-4 flex-shrink-0" />
-                                                  )}
-                                                  <div className="text-left min-w-0 flex-1">
-                                                    <h6 className="font-medium text-sm md:text-base truncate">{getUserName(score.speaker_id)}</h6>
-                                                    <p className="text-xs md:text-sm text-muted-foreground">{score.position}</p>
-                                                  </div>
-                                                </div>
-                                                <div className="text-right flex-shrink-0 ml-2">
-                                                  <div className="text-xl md:text-2xl font-bold text-primary">{score.score}</div>
-                                                  <div className="text-xs md:text-sm text-muted-foreground">out of 30</div>
-                                                </div>
-                                              </Button>
-                                            </CollapsibleTrigger>
-
-                                            <CollapsibleContent>
-                                              <div className="px-3 md:px-4 pb-3 md:pb-4 space-y-3 border-t">
-                                                
-                                                <div className="grid grid-cols-2 lg:grid-cols-4 gap-3 mt-4">
-                                                  {SCORING_CATEGORIES.map((category) => {
-                                                    const CategoryIcon = category.icon;
-                                                    const categoryScore = score[category.key] || 0;
-                                                    return (
-                                                      <div key={category.key} className="text-center">
-                                                        <div className="flex items-center justify-center gap-1 mb-1">
-                                                          <CategoryIcon className={`h-3 w-3 ${category.color}`} />
-                                                          <span className="text-xs font-medium truncate">{category.label.split(' ')[0]}</span>
-                                                        </div>
-                                                        <div className="text-sm font-bold tabular-nums">{categoryScore}</div>
-                                                        <Progress value={(categoryScore / 25) * 100} className="h-1 mt-1" />
-                                                      </div>
-                                                    );
-                                                  })}
-                                                </div>
-
-                                                
-                                                {score.comments && (
-                                                  <div className="pt-3 border-t">
-                                                    <Label className="text-xs font-medium">Judge Feedback:</Label>
-                                                    <p className="text-sm mt-1 p-2 bg-muted rounded break-words">{score.comments}</p>
-                                                  </div>
-                                                )}
-
-                                                
-                                                {score.bias_detected && (
-                                                  <div className="pt-3 border-t">
-                                                    <Alert variant="destructive">
-                                                      <AlertCircle className="h-4 w-4" />
-                                                      <AlertDescription>
-                                                        <p className="font-medium">Bias Detected</p>
-                                                        {score.bias_explanation && (
-                                                          <p className="text-sm mt-1 break-words">{score.bias_explanation}</p>
-                                                        )}
-                                                      </AlertDescription>
-                                                    </Alert>
-                                                  </div>
-                                                )}
-                                              </div>
-                                            </CollapsibleContent>
-                                          </Card>
-                                        </Collapsible>
-                                      ))}
-                                    </div>
-                                  </div>
-                                )}
-                              </div>
-                            );
-                          })()}
-                        </div>
-                      )}
-
-                      
-                      {ballot.notes && (
-                        <div className="pt-3 border-t">
-                          <Label className="text-sm font-medium">Judge Notes:</Label>
-                          <p className="text-sm mt-1 p-3 bg-muted rounded break-words">{ballot.notes}</p>
-                        </div>
-                      )}
-
-                      
-                      <div className="text-xs text-muted-foreground pt-2 border-t flex flex-col sm:flex-row sm:justify-between gap-2">
-                        <span>
-                          Submitted: {ballot.submitted_at ? new Date(ballot.submitted_at).toLocaleString() : "Not submitted"}
-                        </span>
-                        {ballot.submission_state === "submitted" && (
-                          <span className="text-green-600 font-medium">Final Submission</span>
-                        )}
-                      </div>
-                    </div>
-                  </CollapsibleContent>
-                </Card>
-              </Collapsible>
-            ))}
-
-            
-            {filteredBallots.length === 0 && (
-              <div className="text-center py-8 text-muted-foreground">
-                <FileText className="h-8 w-8 mx-auto mb-2 opacity-50" />
-                <p>No ballot details available</p>
-                {selectedJudge !== "all" && (
-                  <p className="text-sm">This judge hasn&#39;t submitted a ballot yet</p>
-                )}
-              </div>
-            )}
-
-            
-            {(debate?.shared_notes?.length ?? 0) > 0 && (
-              <Card>
-                <CardHeader className="pb-3">
-                  <CardTitle className="text-base md:text-lg flex items-center gap-2">
-                    <MessageSquare className="h-4 w-4 md:h-5 md:w-5" />
-                    Shared Notes ({debate.shared_notes?.length ?? 0})
-                  </CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <div className="space-y-3">
-                    {(debate.shared_notes ?? []).map((note: any, idx: number) => (
-                      <div key={idx} className="p-3 bg-muted rounded-lg">
-                        <div className="flex flex-col sm:flex-row sm:justify-between sm:items-start gap-2 mb-2">
-                          <span className="font-medium text-sm">{getUserName(note.author)}</span>
-                          <div className="flex flex-col sm:flex-row sm:items-center gap-2">
-                            <Badge variant="outline" className="text-xs w-fit">
-                              {note.visibility}
-                            </Badge>
-                            <span className="text-xs text-muted-foreground">
-                              {new Date(note.timestamp).toLocaleString()}
-                            </span>
-                          </div>
-                        </div>
-                        <p className="text-sm break-words">{note.content}</p>
-                      </div>
-                    ))}
-                  </div>
-                </CardContent>
-              </Card>
-            )}
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-
-interface FlagBallotDialogProps {
-  debate: EnrichedDebate | null;
-  isOpen: boolean;
-  onClose: () => void;
-  onFlag: (debate: EnrichedDebate, reason: string, ballotIds: string[]) => Promise<void>;
-  userRole: UserRole;
-}
-
-function FlagBallotDialog({ debate, isOpen, onClose, onFlag, userRole }: FlagBallotDialogProps) {
-  const [reason, setReason] = useState("");
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [selectedBallots, setSelectedBallots] = useState<string[]>([]);
-
-  const availableBallots = useMemo(() => {
-    if (userRole === "admin") {
-
-      return debate?.judges_ballots
-        ?.filter((jb: any) => jb.ballot?.submission_state === "submitted")
-        ?.map((jb: any) => ({
-          id: jb.ballot._id,
-          judgeName: jb.judge_name || `Judge ${jb.judge_id.slice(-4)}`,
-          isHeadJudge: debate.head_judge_id === jb.judge_id,
-          isAlreadyFlagged: jb.ballot.flagged ?? false,
-          ballot: jb.ballot
-        })) || [];
-    } else {
-
-      return debate?.my_submission ? [{
-        id: debate.my_submission._id,
-        judgeName: "Your Ballot",
-        isHeadJudge: debate.head_judge_id === debate.my_submission.judge_id,
-        isAlreadyFlagged: debate.my_submission.notes?.includes("[FLAG:") || debate.my_submission.notes?.includes("[JUDGE FLAG:"),
-        ballot: debate.my_submission
-      }] : [];
-    }
-  }, [debate, userRole]);
-
-  useEffect(() => {
-    if (isOpen && availableBallots.length > 0) {
-      if (userRole === "volunteer") {
-        setSelectedBallots([availableBallots[0].id]);
-      } else {
-
-        const unflaggedBallots = availableBallots
-          .filter((ballot: { isAlreadyFlagged: any; }) => !ballot.isAlreadyFlagged)
-          .map((ballot: { id: any; }) => ballot.id);
-        setSelectedBallots(unflaggedBallots);
-      }
-    }
-  }, [isOpen, availableBallots, userRole]);
-
-  const handleBallotToggle = (ballotId: string, checked: boolean) => {
-    if (checked) {
-      setSelectedBallots(prev => [...prev, ballotId]);
-    } else {
-      setSelectedBallots(prev => prev.filter(id => id !== ballotId));
-    }
-  };
-
-  const handleSelectAll = (checked: boolean) => {
-    if (checked) {
-      setSelectedBallots(availableBallots.map((ballot: { id: any; }) => ballot.id));
-    } else {
-      setSelectedBallots([]);
-    }
-  };
-
-  const handleFlag = async () => {
-    if (!reason.trim() || selectedBallots.length === 0) return;
-
-    setIsSubmitting(true);
-    try {
-      if (!debate) return;
-      await onFlag(debate, reason, selectedBallots);
-      setReason("");
-      setSelectedBallots([]);
-      onClose();
-    } catch (error) {
-      console.error("Failed to flag ballot:", error);
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  const flaggableBallots = availableBallots.filter((ballot: { isAlreadyFlagged: any; }) => !ballot.isAlreadyFlagged);
-  const alreadyFlaggedCount = availableBallots.length - flaggableBallots.length;
-
-  return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent className="max-w-md">
-        <DialogHeader>
-          <DialogTitle className="flex items-center gap-2">
-            <Flag className="h-5 w-5" />
-            Flag Ballot - {debate?.room_name}
-          </DialogTitle>
-        </DialogHeader>
-
-        <div className="space-y-4">
-
-          {userRole === "admin" && availableBallots.length > 0 && (
-            <div className="space-y-3">
-              <div className="flex items-center justify-between">
-                <Label className="text-sm font-medium">Select Ballots to Flag:</Label>
-                {flaggableBallots.length > 1 && (
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => handleSelectAll(selectedBallots.length !== flaggableBallots.length)}
-                    className="h-auto p-1 text-xs"
-                  >
-                    {selectedBallots.length === flaggableBallots.length ? "Deselect All" : "Select All"}
-                  </Button>
-                )}
-              </div>
-
-              <div className="space-y-2 max-h-40 overflow-y-auto">
-                {availableBallots.map((ballot: any) => (
-                  <div
-                    key={ballot.id}
-                    className={`flex items-center space-x-3 p-2 border rounded ${
-                      ballot.isAlreadyFlagged ? 'border-red-200 bg-red-50' : 'border-border'
-                    }`}
-                  >
-                    <Checkbox
-                      id={`ballot-${ballot.id}`}
-                      checked={selectedBallots.includes(ballot.id)}
-                      onCheckedChange={(checked) => handleBallotToggle(ballot.id, !!checked)}
-                      disabled={ballot.isAlreadyFlagged}
-                    />
-
-                    <div className="flex-1 min-w-0">
-                      <div className="flex items-center gap-2">
-                        <label
-                          htmlFor={`ballot-${ballot.id}`}
-                          className={`text-sm font-medium cursor-pointer ${
-                            ballot.isAlreadyFlagged ? 'text-muted-foreground' : ''
-                          }`}
-                        >
-                          {ballot.judgeName}
-                        </label>
-                        {ballot.isHeadJudge && (
-                          <Badge variant="default" className="gap-1">
-                            <Crown className="h-3 w-3" />
-                            Head
-                          </Badge>
-                        )}
-                      </div>
-
-                      {ballot.isAlreadyFlagged && (
-                        <div className="text-xs text-red-600 mt-1">
-                          Already flagged
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                ))}
-              </div>
-
-              {alreadyFlaggedCount > 0 && (
-                <div className="text-xs text-muted-foreground">
-                  {alreadyFlaggedCount} ballot{alreadyFlaggedCount > 1 ? 's' : ''} already flagged
-                </div>
-              )}
-
-              {flaggableBallots.length === 0 && (
-                <div className="text-center py-4 text-muted-foreground">
-                  <Flag className="h-8 w-8 mx-auto mb-2 opacity-50" />
-                  <p className="text-sm">All ballots are already flagged</p>
-                </div>
-              )}
-            </div>
-          )}
-
-
-          {userRole === "volunteer" && availableBallots.length > 0 && (
-            <div className="p-3 bg-muted rounded-lg">
-              <div className="flex items-center gap-2">
-                <Flag className="h-4 w-4" />
-                <span className="text-sm font-medium">Flagging your ballot</span>
-              </div>
-              <p className="text-xs text-muted-foreground mt-1">
-                This will flag your ballot for admin review
-              </p>
-            </div>
-          )}
-
-
-          {availableBallots.length === 0 && (
-            <div className="text-center py-4 text-muted-foreground">
-              <AlertTriangle className="h-8 w-8 mx-auto mb-2 opacity-50" />
-              <p className="text-sm">No ballots available to flag</p>
-              <p className="text-xs">
-                {userRole === "admin"
-                  ? "No submitted ballots found for this debate"
-                  : "You haven't submitted a ballot for this debate yet"
-                }
-              </p>
-            </div>
-          )}
-
-          <div>
-            <Label>Reason for flagging:</Label>
-            <Textarea
-              placeholder="Describe the issue with the ballot(s)..."
-              value={reason}
-              onChange={(e) => setReason(e.target.value)}
-              rows={3}
-              disabled={availableBallots.length === 0}
-            />
-          </div>
-
-          <div className="flex gap-2 pt-4">
-            <Button
-              onClick={handleFlag}
-              disabled={
-                !reason.trim() ||
-                selectedBallots.length === 0 ||
-                isSubmitting ||
-                availableBallots.length === 0
-              }
-              variant="destructive"
-              className="flex-1"
-            >
-              {isSubmitting ? (
-                <>
-                  <Loader2 className="h-4 w-4 animate-spin mr-2" />
-                  Flagging...
-                </>
-              ) : (
-                <>
-                  <Flag className="h-4 w-4 mr-2" />
-                  Flag {selectedBallots.length > 1 ? `${selectedBallots.length} Ballots` : 'Ballot'}
-                </>
-              )}
-            </Button>
-            <Button onClick={onClose} variant="outline">
-              Cancel
-            </Button>
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
   );
 }
